Rename and move tempfs to tools/ctl/localfiles
