+++ tools/ctl/localfiles/localfiles.go
@@ -1,4 +1,4 @@
-package tempfs
+package localfiles
 
 import (
 	"os"
